+++ test/components/datagrid/datagrid.e2e-spec.js
@@ -4433,3 +4433,41 @@ describe('Datagrid with select rows across pages tests', () => {
     expect(await element(by.css(row2)).getAttribute('class')).toMatch('is-selected');
   });
 });
+
+describe('Datagrid treegrid Tooltip tests', () => {
+  beforeEach(async () => {
+    await utils.setPage('/components/datagrid/test-tree-tooltip?layout=nofrills');
+
+    const datagridEl = await element(by.css('#datagrid tbody tr:nth-child(1)'));
+    await browser.driver
+      .wait(protractor.ExpectedConditions.presenceOf(datagridEl), config.waitsFor);
+  });
+
+  it('Should not have errors', async () => {
+    await utils.checkForErrors();
+  });
+
+  it('Should show tooltip get data with callback', async () => {
+    await browser.actions().mouseMove(element(by.css('tbody tr[aria-rowindex="10"] td[aria-colindex="4"]'))).perform();
+    await browser.driver
+      .wait(protractor.ExpectedConditions.visibilityOf(await element(by.css('.grid-tooltip'))), config.waitsFor);
+    const tooltip = await element(by.css('.grid-tooltip'));
+
+    expect(await tooltip.getAttribute('class')).not.toContain('is-hidden');
+    expect(await tooltip.getText()).toEqual('Ordered at 7:04 AM This is cell: 3');
+  });
+
+  it('Should show tooltip on text cut off for indented area', async () => {
+    await browser.actions().mouseMove(element(by.css('tbody tr[aria-rowindex="10"] td[aria-colindex="1"]'))).perform();
+    await browser.driver
+      .wait(protractor.ExpectedConditions.visibilityOf(await element(by.css('.grid-tooltip'))), config.waitsFor);
+    const tooltip = await element(by.css('.grid-tooltip'));
+
+    expect(await tooltip.getAttribute('class')).not.toContain('is-hidden');
+    expect(await tooltip.getText()).toEqual('Follow up action with Residental New York');
+    await browser.actions().mouseMove(element(by.css('tbody tr[aria-rowindex="11"] td[aria-colindex="1"]'))).perform();
+    await browser.driver.sleep(config.waitsFor);
+
+    expect(await tooltip.getAttribute('class')).toContain('is-hidden');
+  });
+});